+++ android/src/main/java/com/reactlibrary/rnwifi/ForceWifiUsageErrorCodes.java
@@ -1,0 +1,5 @@
+package com.reactlibrary.rnwifi;
+
+public enum ForceWifiUsageErrorCodes {
+    couldNotGetConnectivityManager,
+}